cleanup: consolidate tab components
